Add files via upload
This datafile is drawn from a larger meta-analysis conducted by the HEDCO Institute The full data are publicly available on OSF, https://osf.io/c7nyz/. This subset of data were selected for training purposes and are not intended to represent the larger body of literature included in the meta-analysis. The cleaning script is hosted in this repository for individuals interested in replicating the subset of included data.